pyproject.toml
[project]
name = "post"
version = "0.1.0"
description = "Multi-platform posting system for Mwmbl project updates"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matrix-nio>=0.24.0",
    "PyGithub>=2.1.1",
    "Mastodon.py>=1.8.1",
    "tweepy>=4.14.0",
    "anthropic>=0.34.0",
    "httpx>=0.27.0",
    "GitPython>=3.1.40",
    "python-decouple>=3.8",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.2",
    "schedule>=1.2.0",
    "click>=8.1.7"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
